+++ src/resources/views/custom_fields/otp_secret/create.blade.php
@@ -1,0 +1,30 @@
+<?php
+/** @var \Yaro\Jarboe\Etc\CustomFields\OtpSecret $field */
+?>
+
+<label class="label">
+    {{ $field->title() }}
+</label>
+
+<label class="input {{ $errors->has($field->name()) ? 'state-error' : '' }}">
+    @if ($field->hasTooltip())
+        <i class="icon-append fa fa-question-circle"></i>
+    @endif
+
+    <input type="text"
+           value="{{ $field->oldOrDefault() }}"
+           readonly
+           placeholder="{{ $field->getPlaceholder() }}">
+
+    @if ($field->hasTooltip())
+        <b class="tooltip tooltip-top-right">
+            <i class="fa fa-warning txt-color-teal"></i>
+            {{ $field->getTooltip() }}
+        </b>
+    @endif
+
+    @foreach ($errors->get($field->name()) as $message)
+        <div class="note note-error">{{ $message }}</div>
+    @endforeach
+
+</label>